утечки исправлены до конца, переписан метод doDefaultCloseAction
