+++ src/main/java/org/example/gui/GameWindow.java
@@ -6,64 +6,19 @@ import javax.swing.*;
 import javax.swing.event.InternalFrameEvent;
 import javax.swing.event.InternalFrameListener;
 
-public class GameWindow extends JInternalFrame
+public class GameWindow extends inter
 {
     private final GameVisualizer m_visualizer;
-    public GameWindow() 
+    public GameWindow()
     {
         super("Игровое поле", true, true, true, true);
+
         m_visualizer = new GameVisualizer();
-        setDefaultCloseOperation(JInternalFrame.DO_NOTHING_ON_CLOSE);
-        addInternalFrameListener(new InternalFrameListener() {
 
-            @Override
-            public void internalFrameOpened(InternalFrameEvent e) {
-
-            }
-
-            @Override
-            public void internalFrameClosing(InternalFrameEvent event) {
-                Object[] options = { "Да", "Нет!" };
-                int n = JOptionPane
-                        .showOptionDialog(event.getInternalFrame(), "Закрыть окно?",
-                                "Подтверждение", JOptionPane.YES_NO_OPTION,
-                                JOptionPane.QUESTION_MESSAGE, null, options,
-                                options[0]);
-                if (n == 0) {
-                    event.getInternalFrame().setVisible(false);
-                    setDefaultCloseOperation(EXIT_ON_CLOSE);
-                }
-            }
-
-            @Override
-            public void internalFrameClosed(InternalFrameEvent e) {
-
-            }
-
-            @Override
-            public void internalFrameIconified(InternalFrameEvent e) {
-
-            }
-
-            @Override
-            public void internalFrameDeiconified(InternalFrameEvent e) {
-
-            }
-
-            @Override
-            public void internalFrameActivated(InternalFrameEvent e) {
-
-            }
-
-            @Override
-            public void internalFrameDeactivated(InternalFrameEvent e) {
-
-            }
-
-        });
         JPanel panel = new JPanel(new BorderLayout());
         panel.add(m_visualizer, BorderLayout.CENTER);
         getContentPane().add(panel);
         pack();
     }
+
 }
