+++ src/main/java/org/example/gui/LogWindow.java
@@ -11,72 +11,27 @@ import log.LogChangeListener;
 import log.LogEntry;
 import log.LogWindowSource;
 
-public class LogWindow extends JInternalFrame implements LogChangeListener
+public class LogWindow extends inter implements LogChangeListener
 {
     private LogWindowSource m_logSource;
     private TextArea m_logContent;
 
-    public LogWindow(LogWindowSource logSource) 
+    public LogWindow(LogWindowSource logSource)
     {
         super("Протокол работы", true, true, true, true);
         m_logSource = logSource;
         m_logSource.registerListener(this);
         m_logContent = new TextArea("");
         m_logContent.setSize(200, 500);
-        setDefaultCloseOperation(JInternalFrame.DO_NOTHING_ON_CLOSE);
-        addInternalFrameListener(new InternalFrameListener() {
 
-            @Override
-            public void internalFrameOpened(InternalFrameEvent e) {
-
-            }
-
-            @Override
-            public void internalFrameClosing(InternalFrameEvent event) {
-                Object[] options = { "Да", "Нет!" };
-                int n = JOptionPane
-                        .showOptionDialog(event.getInternalFrame(), "Закрыть окно?",
-                                "Подтверждение", JOptionPane.YES_NO_OPTION,
-                                JOptionPane.QUESTION_MESSAGE, null, options,
-                                options[0]);
-                if (n == 0) {
-                    event.getInternalFrame().setVisible(false);
-                    setDefaultCloseOperation(EXIT_ON_CLOSE);
-                }
-            }
-
-            @Override
-            public void internalFrameClosed(InternalFrameEvent e) {
-
-            }
-
-            @Override
-            public void internalFrameIconified(InternalFrameEvent e) {
-
-            }
-
-            @Override
-            public void internalFrameDeiconified(InternalFrameEvent e) {
-
-            }
-
-            @Override
-            public void internalFrameActivated(InternalFrameEvent e) {
-
-            }
-
-            @Override
-            public void internalFrameDeactivated(InternalFrameEvent e) {
-
-            }
-
-        });
         JPanel panel = new JPanel(new BorderLayout());
         panel.add(m_logContent, BorderLayout.CENTER);
         getContentPane().add(panel);
         pack();
         updateLogContent();
     }
+
+
 
     private void updateLogContent()
     {
@@ -88,7 +43,13 @@ public class LogWindow extends JInternalFrame implements LogChangeListener
         m_logContent.setText(content.toString());
         m_logContent.invalidate();
     }
-    
+
+    @Override
+    public void doDefaultCloseAction() {
+        m_logSource.unregisterListener(this);
+        super.doDefaultCloseAction();
+    }
+
     @Override
     public void onLogChanged()
     {
