+++ src/main/java/org/example/log/LogWindowSource.java
@@ -14,12 +14,18 @@ import java.util.concurrent.ArrayBlockingQueue;
  * ограниченного размера)
  *
  * изменения:
- * Используется ArrayBlockingQueue для хранения сообщений лога, что позволяет избежать утечек ресурсов и ограничить размер лога величиной m_iQueueLength.
- * Избавлены от необходимости хранить все слушатели в виде массива, который нужно было пересоздавать каждый раз при добавлении или удалении слушателя. Вместо этого используется обычный список m_listeners, синхронизированный на запись/удаление.
- * Используется offer() вместо add() при добавлении сообщения в очередь, чтобы избежать исключения в случае, если очередь заполнена.
+ * Используется ArrayBlockingQueue для хранения сообщений лога, что позволяет избежать утечек ресурсов и ограничить
+ * размер лога величиной m_iQueueLength.
+ * Избавлены от необходимости хранить все слушатели в виде массива, который нужно было пересоздавать каждый раз при
+ * добавлении или удалении слушателя. Вместо этого используется обычный список m_listeners, синхронизированный
+ * на запись/удаление.
+ * Используется offer() вместо add() при добавлении сообщения в очередь, чтобы избежать исключения в случае,
+ * если очередь заполнена.
  * Метод size() теперь возвращает размер m_queue.
- * Метод range() теперь проверяет, что count неотрицательное число и возвращает пустой список, если startFrom выходит за пределы очереди или count равен нулю.
- * Метод range() теперь использует метод ArrayList.subList() непосредственно над m_queue, а не создает новый список, что улучшает производительность и уменьшает использование памяти.
+ * Метод range() теперь проверяет, что count неотрицательное число и возвращает пустой список, если startFrom выходит
+ * за пределы очереди или count равен нулю.
+ * Метод range() теперь использует метод ArrayList.subList() непосредственно над m_queue, а не создает новый список,
+ * что улучшает производительность и уменьшает использование памяти.
  * Метод all() теперь создает новый список из m_queue, вместо создания списка итератором.
  */
 public class LogWindowSource {
