+++ src/main/resources/resources_en_US.properties
@@ -1,0 +1,1 @@
+Language = Language
