+++ src/main/resources/resources_ru_RU.properties
@@ -1,0 +1,1 @@
+Language = ttt
